Modified the sketch for the classification on Arduino
I have done a lot of changes. I changed the model trained, so you have to replace the folder of the neural network inside the Arduino Library with the new folder I put in the commit. I made changes to get it work and now we are almost done. I can see the NN working and giving inferences about the data from accelerometer and gyroscope.